+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Unreleased
+- Added Message Template API.
+- Added Currency and Datetime resources.
+- Added Media resource.
+- Added Component and ParameterObject resource.
+- Fixed bug in recipient_number in Messages API.
 
 # v 0.0.2
 - Implement read message.
 - Implement Yard doc.
-